requirements.txt
async-generator==1.10
attrs==22.1.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
distlib==0.3.5
filelock==3.7.1
h11==0.13.0
idna==3.3
joblib==1.1.0
mutagen==1.45.1
nltk==3.7
outcome==1.2.0
platformdirs==2.5.2
praw==7.6.0
prawcore==2.3.0
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dotenv==0.20.0
regex==2022.7.25
requests==2.28.1
selenium==4.3.0
sniffio==1.2.0
sortedcontainers==2.4.0
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
update-checker==0.18.0
urllib3==1.26.11
virtualenv==20.16.0
websocket-client==1.3.3
wsproto==1.1.0
